EDU-1235 Move room media logic to room controller and service
