+++ src/server/controller-factory.js
@@ -9,7 +9,6 @@ import AdminController from './admin-controller.js';
 import BatchController from './batch-controller.js';
 import SearchController from './search-controller.js';
 import StaticController from './static-controller.js';
-import StorageController from './storage-controller.js';
 import CommentController from './comment-controller.js';
 import SettingsController from './settings-controller.js';
 import DocumentController from './document-controller.js';
@@ -18,6 +17,7 @@ import PdfJsApiController from './pdfjs-api-controller.js';
 import DashboardController from './dashboard-controller.js';
 import RedactionController from './redaction-controller.js';
 import UserAgentController from './user-agent-controller.js';
+import StoragePlanController from './storage-plan-controller.js';
 import NotificationController from './notification-controller.js';
 import MediaLibraryController from './media-library-controller.js';
 
@@ -34,7 +34,7 @@ const pageAndApiControllers = [
   UserController,
   DocumentController,
   MediaLibraryController,
-  StorageController,
+  StoragePlanController,
   BatchController,
   RoomController,
   DashboardController,
