+++ src/server/room-controller.js
@@ -1,3 +1,5 @@
+import os from 'node:os';
+import multer from 'multer';
 import express from 'express';
 import httpErrors from 'http-errors';
 import routes from '../utils/routes.js';
@@ -9,12 +11,12 @@ import RoomService from '../services/room-service.js';
 import UserService from '../services/user-service.js';
 import MailService from '../services/mail-service.js';
 import ServerConfig from '../bootstrap/server-config.js';
-import StorageService from '../services/storage-service.js';
 import DocumentService from '../services/document-service.js';
 import needsPermission from '../domain/needs-permission-middleware.js';
 import needsAuthentication from '../domain/needs-authentication-middleware.js';
 import ClientDataMappingService from '../services/client-data-mapping-service.js';
-import { validateBody, validateParams, validateQuery } from '../domain/validation-middleware.js';
+import uploadLimitExceededMiddleware from '../domain/upload-limit-exceeded-middleware.js';
+import { validateBody, validateFile, validateParams, validateQuery } from '../domain/validation-middleware.js';
 import { isRoomOwner, isRoomOwnerOrInvitedCollaborator, isRoomOwnerOrInvitedMember } from '../utils/room-utils.js';
 import {
   NOT_ROOM_OWNER_ERROR_MESSAGE,
@@ -39,24 +41,54 @@ import {
   deleteRoomInvitationParamsSchema,
   postRoomInvitationConfirmBodySchema,
   getAuthorizeResourcesAccessParamsSchema,
-  getRoomMembershipConfirmationParamsSchema
+  getRoomMembershipConfirmationParamsSchema,
+  getAllOrPostRoomMediaParamsSchema,
+  deleteRoomMediaParamsSchema
 } from '../domain/schemas/room-schemas.js';
 
 const jsonParser = express.json();
+const multipartParser = multer({ dest: os.tmpdir() });
+
 const { NotFound, Forbidden, Unauthorized, BadRequest } = httpErrors;
 
 export default class RoomController {
-  static dependencies = [ServerConfig, RoomService, DocumentService, UserService, StorageService, MailService, ClientDataMappingService, PageRenderer];
+  static dependencies = [ServerConfig, RoomService, DocumentService, UserService, MailService, ClientDataMappingService, PageRenderer];
 
-  constructor(serverConfig, roomService, documentService, userService, storageService, mailService, clientDataMappingService, pageRenderer) {
+  constructor(serverConfig, roomService, documentService, userService, mailService, clientDataMappingService, pageRenderer) {
     this.roomService = roomService;
     this.userService = userService;
     this.mailService = mailService;
     this.serverConfig = serverConfig;
     this.pageRenderer = pageRenderer;
     this.documentService = documentService;
-    this.storageService = storageService;
     this.clientDataMappingService = clientDataMappingService;
+  }
+
+  async handleGetRoomMediaOverview(req, res) {
+    const { user } = req;
+    const roomMediaOverview = await this.roomService.getRoomMediaOverview({ user });
+    return res.send(roomMediaOverview);
+  }
+
+  async handleGetAllRoomMedia(req, res) {
+    const { user } = req;
+    const { roomId } = req.params;
+    const roomMedia = await this.roomService.getAllRoomMedia({ user, roomId });
+    return res.send(roomMedia);
+  }
+
+  async handlePostRoomMedia(req, res) {
+    const { user, file } = req;
+    const { roomId } = req.params;
+    const roomMedia = await this.roomService.createRoomMedia({ user, roomId, file });
+    return res.status(201).send(roomMedia);
+  }
+
+  async handleDeleteRoomMedia(req, res) {
+    const { user } = req;
+    const { roomId, name } = req.params;
+    const roomMedia = await this.roomService.deleteRoomMedia({ user, roomId, name });
+    return res.send(roomMedia);
   }
 
   async handleGetRoomMembershipConfirmationPage(req, res) {
@@ -167,13 +199,13 @@ export default class RoomController {
     const roomOwner = await this.userService.getUserById(ownerId);
 
     if (!roomOwner) {
-      throw new BadRequest(`Unknown room owner with ID '${ownerId}'`);
+      throw new NotFound(`Unknown room owner with ID '${ownerId}'`);
     }
 
     const rooms = await this.roomService.getRoomsOwnedByUser(ownerId);
 
     for (const room of rooms) {
-      await this._deleteRoom({ room, roomOwner });
+      await this._deleteRoomAndNotify({ room, roomOwner });
     }
 
     return res.send({});
@@ -193,7 +225,7 @@ export default class RoomController {
       throw new Forbidden(NOT_ROOM_OWNER_ERROR_MESSAGE);
     }
 
-    await this._deleteRoom({ room, roomOwner: user });
+    await this._deleteRoomAndNotify({ room, roomOwner: user });
 
     return res.send({});
   }
@@ -346,8 +378,8 @@ export default class RoomController {
     return res.status(200).end();
   }
 
-  async _deleteRoom({ room, roomOwner }) {
-    await this.storageService.deleteRoomAndResources({ roomId: room._id, roomOwnerId: roomOwner._id });
+  async _deleteRoomAndNotify({ room, roomOwner }) {
+    await this.roomService.deleteRoom({ room, roomOwner });
     await this.mailService.sendRoomDeletionNotificationEmails({
       roomName: room.name,
       ownerName: roomOwner.displayName,
@@ -404,6 +436,36 @@ export default class RoomController {
       (req, res) => this.handleDeleteRoomMember(req, res)
     );
 
+    router.get(
+      '/api/v1/room-media-overview',
+      needsPermission(permissions.BROWSE_STORAGE),
+      (req, res) => this.handleGetRoomMediaOverview(req, res)
+    );
+
+    router.get(
+      '/api/v1/room-media/:roomId',
+      needsPermission(permissions.BROWSE_STORAGE),
+      validateParams(getAllOrPostRoomMediaParamsSchema),
+      (req, res) => this.handleGetAllRoomMedia(req, res)
+    );
+
+    router.post(
+      '/api/v1/room-media/:roomId',
+      needsPermission(permissions.CREATE_CONTENT),
+      uploadLimitExceededMiddleware(),
+      multipartParser.single('file'),
+      validateFile('file'),
+      validateParams(getAllOrPostRoomMediaParamsSchema),
+      (req, res) => this.handlePostRoomMedia(req, res)
+    );
+
+    router.delete(
+      '/api/v1/room-media/:roomId/:name',
+      needsPermission(permissions.DELETE_OWN_PRIVATE_CONTENT),
+      validateParams(deleteRoomMediaParamsSchema),
+      (req, res) => this.handleDeleteRoomMedia(req, res)
+    );
+
     router.post(
       '/api/v1/room-invitations',
       [needsPermission(permissions.CREATE_CONTENT), jsonParser, validateBody(postRoomInvitationsBodySchema)],
@@ -423,7 +485,7 @@ export default class RoomController {
     );
 
     router.post(
-      '/api/v1/rooms/:roomId/message',
+      '/api/v1/rooms/:roomId/messages',
       [needsPermission(permissions.CREATE_CONTENT), jsonParser, validateBody(postRoomMessageBodySchema)],
       (req, res) => this.handlePostRoomMessage(req, res)
     );
