+++ src/services/room-service.js
@@ -1,35 +1,68 @@
 import moment from 'moment';
 import httpErrors from 'http-errors';
+import prettyBytes from 'pretty-bytes';
 import Logger from '../common/logger.js';
 import Cdn from '../repositories/cdn.js';
 import uniqueId from '../utils/unique-id.js';
+import urlUtils from '../utils/url-utils.js';
 import cloneDeep from '../utils/clone-deep.js';
 import RoomStore from '../stores/room-store.js';
 import LockStore from '../stores/lock-store.js';
 import UserStore from '../stores/user-store.js';
+import CommentStore from '../stores/comment-store.js';
 import DocumentStore from '../stores/document-store.js';
+import StoragePlanStore from '../stores/storage-plan-store.js';
 import TransactionRunner from '../stores/transaction-runner.js';
-import { getRoomMediaRoomPath } from '../utils/storage-utils.js';
 import RoomInvitationStore from '../stores/room-invitation-store.js';
+import DocumentRevisionStore from '../stores/document-revision-store.js';
 import { ensureIsExcluded, getSymmetricalDifference } from '../utils/array-utils.js';
+import { createUniqueStorageFileName, getRoomMediaRoomPath } from '../utils/storage-utils.js';
+import { isRoomOwner, isRoomOwnerOrInvitedCollaborator, isRoomOwnerOrInvitedMember } from '../utils/room-utils.js';
 import { INVALID_ROOM_INVITATION_REASON, PENDING_ROOM_INVITATION_EXPIRATION_IN_DAYS } from '../domain/constants.js';
 
-const { BadRequest, NotFound } = httpErrors;
+const { BadRequest, Forbidden, NotFound } = httpErrors;
 
 const logger = new Logger(import.meta.url);
 
 export default class RoomService {
-  static dependencies = [Cdn, RoomStore, RoomInvitationStore, LockStore, UserStore, DocumentStore, TransactionRunner];
+  static dependencies = [
+    Cdn,
+    RoomStore,
+    RoomInvitationStore,
+    DocumentRevisionStore,
+    DocumentStore,
+    CommentStore,
+    UserStore,
+    StoragePlanStore,
+    LockStore,
+    TransactionRunner
+  ];
 
-  constructor(cdn, roomStore, roomInvitationStore, lockStore, userStore, documentStore, transactionRunner) {
+  constructor(
+    cdn,
+    roomStore,
+    roomInvitationStore,
+    documentRevisionStore,
+    documentStore,
+    commentStore,
+    userStore,
+    storagePlanStore,
+    lockStore,
+    transactionRunner
+  ) {
     this.cdn = cdn;
     this.roomStore = roomStore;
     this.lockStore = lockStore;
     this.userStore = userStore;
+    this.commentStore = commentStore;
     this.documentStore = documentStore;
+    this.storagePlanStore = storagePlanStore;
     this.transactionRunner = transactionRunner;
     this.roomInvitationStore = roomInvitationStore;
+    this.documentRevisionStore = documentRevisionStore;
   }
+
+  // ROOMS
 
   getRoomById(roomId) {
     return this.roomStore.getRoomById(roomId);
@@ -49,15 +82,6 @@ export default class RoomService {
 
   getRoomsByOwnerOrCollaboratorUser(userId) {
     return this.roomStore.getRoomsByOwnerOrCollaboratorUser({ userId });
-  }
-
-  getRoomInvitationsByEmail(email) {
-    return this.roomInvitationStore.getRoomInvitationsByEmail(email);
-  }
-
-  async isRoomOwnerOrMember(roomId, userId) {
-    const room = await this.roomStore.getRoomsByIdOwnedOrJoinedByUser({ roomId, userId });
-    return !!room;
   }
 
   async createRoom({ name, slug, isCollaborative, user }) {
@@ -91,67 +115,156 @@ export default class RoomService {
     return newRoom;
   }
 
-  async createRoomMessage({ room, text, emailNotification }) {
-    const messages = cloneDeep(room.messages);
-    messages.push({
-      key: uniqueId.create(),
-      text,
-      emailNotification,
-      createdOn: new Date()
-    });
-
-    await this.roomStore.updateRoomMessages(room._id, messages);
-
-    const updatedRoom = await this.roomStore.getRoomById(room._id);
-
-    return updatedRoom;
-  }
-
-  async deleteRoomMessage({ room, messageKey }) {
-    const message = room.messages.find(m => m.key === messageKey);
-    const remainingMessages = ensureIsExcluded(room.messages, message);
-    await this.roomStore.updateRoomMessages(room._id, remainingMessages);
-
-    const updatedRoom = await this.roomStore.getRoomById(room._id);
-
-    return updatedRoom;
-  }
-
-  async updateRoomMetadata(roomId, { name, slug, isCollaborative, description }) {
-    await this.roomStore.updateRoomMetadata(
-      roomId,
-      {
-        name: name.trim(),
-        slug: (slug || '').trim(),
-        isCollaborative,
-        description: (description || '').trim(),
-        updatedOn: new Date()
-      }
-    );
-    const updatedRoom = await this.roomStore.getRoomById(roomId);
-
-    return updatedRoom;
-  }
-
-  async updateRoomDocumentsOrder(roomId, documentIds) {
+  async deleteRoom({ room, roomOwner }) {
     let lock;
 
     try {
-      lock = await this.lockStore.takeRoomLock(roomId);
-      const room = await this.roomStore.getRoomById(roomId);
-      const allDocumentIdsCorrectlyProvided = getSymmetricalDifference(documentIds, room.documents).length === 0;
+      lock = await this.lockStore.takeUserLock(roomOwner._id);
+      logger.info(`Deleting room with ID ${room._id}`);
 
-      if (!allDocumentIdsCorrectlyProvided) {
-        throw new BadRequest('Incorrect list of document ids was provided.');
-      }
+      await this.transactionRunner.run(async session => {
+        const documents = await this.documentStore.getDocumentsMetadataByRoomId(room._id, { session });
+        const documentIds = documents.map(doc => doc._id);
 
-      await this.roomStore.updateRoomDocuments(roomId, documentIds);
+        await this.commentStore.deleteCommentsByDocumentIds(documentIds, { session });
+        await this.documentRevisionStore.deleteDocumentsByRoomId(room._id, { session });
+        await this.documentStore.deleteDocumentsByRoomId(room._id, { session });
+        await this.roomInvitationStore.deleteRoomInvitationsByRoomId(room._id, { session });
+        await this.roomStore.deleteRoomById(room._id, { session });
+      });
+
+      await this.cdn.deleteDirectory({ directoryPath: getRoomMediaRoomPath(room._id) });
+
+      const { usedBytes } = await this.getRoomMediaOverview({ user: roomOwner });
+      const updatedUser = await this.userStore.updateUserUsedBytes({ userId: roomOwner._id, usedBytes });
+
+      Object.assign(roomOwner, updatedUser);
     } finally {
       await this.lockStore.releaseLock(lock);
     }
+  }
 
-    const updatedRoom = await this.roomStore.getRoomById(roomId);
-    return updatedRoom;
+  // ROOM MEDIA
+
+  async getRoomMediaOverview({ user }) {
+    const storagePlan = user.storage.planId
+      ? await this.storagePlanStore.getStoragePlanById(user.storage.planId)
+      : null;
+
+    const rooms = await this.roomStore.getRoomsByOwnerId(user._id);
+    const objectsPerRoom = await Promise.all(rooms.map(async room => {
+      const objects = await this.cdn.listObjects({ directoryPath: getRoomMediaRoomPath(room._id) });
+      return { room, objects, usedBytes: objects.reduce((accu, obj) => accu + obj.size, 0) };
+    }));
+
+    const usedBytesInAllRooms = objectsPerRoom.reduce((accu, { usedBytes }) => accu + usedBytes, 0);
+
+    return {
+      storagePlan: storagePlan || null,
+      usedBytes: usedBytesInAllRooms,
+      roomStorageList: objectsPerRoom.map(({ room, objects }) => ({
+        roomId: room._id,
+        roomName: room.name,
+        objects
+      }))
+    };
+  }
+
+  async getAllRoomMedia({ user, roomId }) {
+    const room = await this.roomStore.getRoomById(roomId);
+    if (!isRoomOwnerOrInvitedMember({ room, userId: user._id })) {
+      throw new Forbidden(`User is not authorized to access media for room '${roomId}'`);
+    }
+
+    const roomOwner = isRoomOwner({ room, userId: user._id }) ? user : await this.userStore.findActiveUserById(room.owner);
+    const overview = await this.getRoomMediaOverview({ user: roomOwner });
+
+    return {
+      storagePlan: overview.storagePlan,
+      usedBytes: overview.usedBytes,
+      roomStorage: overview.roomStorageList.find(roomStorage => roomStorage.roomId === roomId)
+    };
+  }
+
+  async createRoomMedia({ user, roomId, file }) {
+    let lock;
+    try {
+      const room = await this.roomStore.getRoomById(roomId);
+      if (!isRoomOwnerOrInvitedCollaborator({ room, userId: user._id })) {
+        throw new Forbidden(`User is not authorized to create media for room '${roomId}'`);
+      }
+
+      lock = await this.lockStore.takeUserLock(room.owner);
+
+      const roomOwner = isRoomOwner({ room, userId: user._id }) ? user : await this.userStore.findActiveUserById(room.owner);
+      if (!roomOwner.storage.planId) {
+        throw new BadRequest('Cannot upload to room-media storage without a storage plan');
+      }
+
+      const storagePlan = await this.storagePlanStore.getStoragePlanById(roomOwner.storage.planId);
+      const availableBytes = storagePlan.maxBytes - roomOwner.storage.usedBytes;
+      if (availableBytes < file.size) {
+        throw new BadRequest(`Not enough storage space: available ${prettyBytes(availableBytes)}, required ${prettyBytes(file.size)}`);
+      }
+
+      const parentPath = getRoomMediaRoomPath(roomId);
+      const uniqueFileName = createUniqueStorageFileName(file.originalname);
+      const cdnObjectPath = urlUtils.concatParts(parentPath, uniqueFileName);
+      await this.cdn.uploadObject(cdnObjectPath, file.path);
+
+      const overview = await this.getRoomMediaOverview({ user: roomOwner });
+      const updatedUser = await this.userStore.updateUserUsedBytes({ userId: roomOwner._id, usedBytes: overview.usedBytes });
+
+      Object.assign(roomOwner, updatedUser);
+
+      return {
+        storagePlan: overview.storagePlan,
+        usedBytes: overview.usedBytes,
+        roomStorage: overview.roomStorageList.find(roomStorage => roomStorage.roomId === roomId)
+      };
+    } finally {
+      await this.lockStore.releaseLock(lock);
+    }
+  }
+
+  async deleteRoomMedia({ user, roomId, name }) {
+    let lock;
+    try {
+      lock = await this.lockStore.takeUserLock(user._id);
+
+      const room = await this.roomStore.getRoomById(roomId);
+      if (!isRoomOwner({ room, userId: user._id })) {
+        throw new Forbidden(`User is not authorized to delete media for room '${roomId}'`);
+      }
+
+      const objects = await this.cdn.listObjects({ directoryPath: getRoomMediaRoomPath(roomId) });
+
+      const itemToDelete = objects.find(obj => obj.name === name);
+      if (!itemToDelete) {
+        throw new NotFound(`Object '${name}' could not be found`);
+      }
+
+      await this.cdn.deleteObject(itemToDelete.path);
+
+      const overview = await this.getRoomMediaOverview({ user });
+      const updatedUser = await this.userStore.updateUserUsedBytes({ userId: user._id, usedBytes: overview.usedBytes });
+
+      Object.assign(user, updatedUser);
+
+      return {
+        storagePlan: overview.storagePlan,
+        usedBytes: overview.usedBytes,
+        roomStorage: overview.roomStorageList.find(roomStorage => roomStorage.roomId === roomId)
+      };
+    } finally {
+      await this.lockStore.releaseLock(lock);
+    }
+  }
+
+  // ROOM INVITATIONS
+
+  getRoomInvitationsByEmail(email) {
+    return this.roomInvitationStore.getRoomInvitationsByEmail(email);
   }
 
   getRoomInvitations(roomId) {
@@ -260,6 +373,64 @@ export default class RoomService {
     });
   }
 
+  async deleteRoomInvitation({ room, invitation }) {
+    await this.roomInvitationStore.deleteRoomInvitationById(invitation._id);
+    const remainingRoomInvitations = await this.getRoomInvitations(room._id);
+
+    return remainingRoomInvitations;
+  }
+
+  // ROOM MESSAGES
+
+  async createRoomMessage({ room, text, emailNotification }) {
+    const messages = cloneDeep(room.messages);
+    messages.push({
+      key: uniqueId.create(),
+      text,
+      emailNotification,
+      createdOn: new Date()
+    });
+
+    await this.roomStore.updateRoomMessages(room._id, messages);
+
+    const updatedRoom = await this.roomStore.getRoomById(room._id);
+
+    return updatedRoom;
+  }
+
+  async updateRoomMetadata(roomId, { name, slug, isCollaborative, description }) {
+    await this.roomStore.updateRoomMetadata(
+      roomId,
+      {
+        name: name.trim(),
+        slug: (slug || '').trim(),
+        isCollaborative,
+        description: (description || '').trim(),
+        updatedOn: new Date()
+      }
+    );
+    const updatedRoom = await this.roomStore.getRoomById(roomId);
+
+    return updatedRoom;
+  }
+
+  async deleteRoomMessage({ room, messageKey }) {
+    const message = room.messages.find(m => m.key === messageKey);
+    const remainingMessages = ensureIsExcluded(room.messages, message);
+    await this.roomStore.updateRoomMessages(room._id, remainingMessages);
+
+    const updatedRoom = await this.roomStore.getRoomById(room._id);
+
+    return updatedRoom;
+  }
+
+  // ROOM MEMBERS
+
+  async isRoomOwnerOrMember(roomId, userId) {
+    const room = await this.roomStore.getRoomsByIdOwnedOrJoinedByUser({ roomId, userId });
+    return !!room;
+  }
+
   async removeRoomMember({ room, memberUserId }) {
     const member = room.members.find(m => m.userId === memberUserId);
     const remainingMembers = ensureIsExcluded(room.members, member);
@@ -274,10 +445,26 @@ export default class RoomService {
     await this.roomStore.deleteRoomsMemberById(memberUserId);
   }
 
-  async deleteRoomInvitation({ room, invitation }) {
-    await this.roomInvitationStore.deleteRoomInvitationById(invitation._id);
-    const remainingRoomInvitations = await this.getRoomInvitations(room._id);
+  // ROOM DOCUMENT ORDER
 
-    return remainingRoomInvitations;
+  async updateRoomDocumentsOrder(roomId, documentIds) {
+    let lock;
+
+    try {
+      lock = await this.lockStore.takeRoomLock(roomId);
+      const room = await this.roomStore.getRoomById(roomId);
+      const allDocumentIdsCorrectlyProvided = getSymmetricalDifference(documentIds, room.documents).length === 0;
+
+      if (!allDocumentIdsCorrectlyProvided) {
+        throw new BadRequest('Incorrect list of document ids was provided.');
+      }
+
+      await this.roomStore.updateRoomDocuments(roomId, documentIds);
+    } finally {
+      await this.lockStore.releaseLock(lock);
+    }
+
+    const updatedRoom = await this.roomStore.getRoomById(roomId);
+    return updatedRoom;
   }
 }
