+++ src/services/storage-plan-service.js
@@ -1,0 +1,78 @@
+import httpErrors from 'http-errors';
+import uniqueId from '../utils/unique-id.js';
+import UserStore from '../stores/user-store.js';
+import LockStore from '../stores/lock-store.js';
+import StoragePlanStore from '../stores/storage-plan-store.js';
+
+const { BadRequest } = httpErrors;
+
+export default class StoragePlanService {
+  static dependencies = [StoragePlanStore, UserStore, LockStore];
+
+  constructor(storagePlanStore, userStore, lockStore) {
+    this.lockStore = lockStore;
+    this.userStore = userStore;
+    this.storagePlanStore = storagePlanStore;
+  }
+
+  getAllStoragePlans() {
+    return this.storagePlanStore.getAllStoragePlans();
+  }
+
+  getAllStoragePlansWithAssignedUserCount() {
+    return this.storagePlanStore.getAllStoragePlansWithAssignedUserCount();
+  }
+
+  getStoragePlanById(id) {
+    return this.storagePlanStore.getStoragePlanById(id);
+  }
+
+  async createStoragePlan({ name, maxBytes }) {
+    const storagePlan = {
+      _id: uniqueId.create(),
+      name,
+      maxBytes
+    };
+
+    await this.storagePlanStore.saveStoragePlan(storagePlan);
+    return storagePlan;
+  }
+
+  async updateStoragePlan(storagePlanId, { name, maxBytes }) {
+    const updatedStoragePlan = {
+      _id: storagePlanId,
+      name,
+      maxBytes
+    };
+
+    let lock;
+    try {
+      lock = await this.lockStore.takeStoragePlanLock(storagePlanId);
+      await this.storagePlanStore.saveStoragePlan(updatedStoragePlan);
+    } finally {
+      if (lock) {
+        await this.lockStore.releaseLock(lock);
+      }
+    }
+
+    return updatedStoragePlan;
+  }
+
+  async deleteStoragePlanById(storagePlanId) {
+    let lock;
+
+    try {
+      lock = await this.lockStore.takeStoragePlanLock(storagePlanId);
+      const usersExist = await this.userStore.checkUsersWithStoragePlanExistByStoragePlanId(storagePlanId);
+      if (usersExist) {
+        throw new BadRequest('Storage plan cannot be deleted because it is still in use');
+      }
+
+      await this.storagePlanStore.deleteStoragePlanById(storagePlanId);
+    } finally {
+      if (lock) {
+        await this.lockStore.releaseLock(lock);
+      }
+    }
+  }
+}
